CFO_CA-main/requirements.txt
amqp==5.0.2
appdirs==1.4.4
autopep8==1.5.4
billiard==3.6.3.0
black==20.8b1
celery==5.0.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==2.2
django-celery-beat==2.1.0
django-colorfield==0.1.15
django-cors-headers==3.4.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-environ==0.4.5
django-rest-framework==0.1.0
django-timezone-field==4.0
djangorestframework==3.11.1
filelock==3.0.12
geoip2==2.9.0
getip2==1.6
idna==2.8
importlib-metadata==1.7.0
importlib-resources==3.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.10.1
kombu==5.0.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==20.4
pathspec==0.8.1
# Pillow==6.0.0
prompt-toolkit==3.0.10
# psycopg2==2.8.2
pycodestyle==2.6.0
PyMySQL==0.9.3
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.1
redis==3.5.3
regex==2020.11.13
requests==2.21.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.24.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.1.0
django-queryset-csv==1.1.0
croniter==1.0.15